pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flightforms"
version = "0.1.0"
description = "GA customs form generator API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "authlib>=1.3",
    "PyJWT>=2.8",
    "SQLAlchemy>=2.0",
    "alembic>=1.13",
    "pymysql>=1.1",
    "python-dotenv>=1.0",
    "pypdf>=4.0",
    "python-docx>=1.0",
    "openpyxl>=3.1",
    "euro-aip>=0.2.0",
    "itsdangerous>=2.1",
    "pytz>=2024.1",
]

[project.scripts]
flightforms = "flightforms.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
